+++ src/material/icon/testing/BUILD.bazel
@@ -24,8 +24,8 @@ filegroup(
 )
 
 ng_test_library(
-    name = "harness_tests_lib",
-    srcs = ["shared.spec.ts"],
+    name = "unit_tests_lib",
+    srcs = glob(["**/*.spec.ts"]),
     deps = [
         ":testing",
         "//src/cdk/platform",
@@ -33,22 +33,6 @@ ng_test_library(
         "//src/cdk/testing/testbed",
         "//src/material/icon",
         "@npm//@angular/platform-browser",
-    ],
-)
-
-ng_test_library(
-    name = "unit_tests_lib",
-    srcs = glob(
-        ["**/*.spec.ts"],
-        exclude = [
-            "**/*.e2e.spec.ts",
-            "shared.spec.ts",
-        ],
-    ),
-    deps = [
-        ":harness_tests_lib",
-        ":testing",
-        "//src/material/icon",
     ],
 )
